# web/src/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionPlan": {
        "additionalProperties": false,
        "description": "The plan of section 9.6. The plan hash (A.4) is taken over `model_dump(mode=\"json\")`, so\nequal plans hold every list in one stable order.",
        "properties": {
          "catalogue_questions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Catalogue Questions",
            "type": "array"
          },
          "declines_on_every_branch": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RuleTrace"
            },
            "title": "Declines On Every Branch",
            "type": "array"
          },
          "effects": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlannedEffect"
            },
            "title": "Effects",
            "type": "array"
          },
          "not_evaluated": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/NotEvaluatedNote"
            },
            "title": "Not Evaluated",
            "type": "array"
          },
          "open_choices": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OpenChoice"
            },
            "title": "Open Choices",
            "type": "array"
          },
          "proposed_decline": {
            "default": false,
            "title": "Proposed Decline",
            "type": "boolean"
          },
          "undecided": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UndecidedPage"
            },
            "title": "Undecided",
            "type": "array"
          },
          "underwriter_decline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Underwriter Decline"
          }
        },
        "required": [
          "effects",
          "declines_on_every_branch",
          "proposed_decline",
          "underwriter_decline",
          "undecided",
          "open_choices",
          "catalogue_questions",
          "not_evaluated"
        ],
        "title": "ActionPlan",
        "type": "object"
      },
...
      "BlockerView": {
        "additionalProperties": false,
        "description": "An open blocker. `item_id` is `blockers.id`, the id an `approve` or `reject` names.\n\nThe detail pane offers the actions for every open item (section 11), so the blocker carries\nwhat its action needs: a pending observation's value (`observation`, required exactly for the\nitem kind `observation`). The item kind and, for a review, the cause are in `detail`.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/BlockerDetail"
          },
          "item_id": {
            "title": "Item Id",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "delivery_unknown",
              "underwriter_question",
              "underwriter_review",
              "data",
              "producer_reply"
            ],
            "title": "Kind",
            "type": "string"
          },
          "observation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FactView"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner": {
            "enum": [
              "underwriter",
              "producer",
              "data_team"
            ],
            "title": "Owner",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "kind",
          "owner",
          "detail",
          "observation"
        ],
        "title": "BlockerView",
        "type": "object"
      },
      "ChatRequest": {
        "additionalProperties": false,
        "description": "One chat turn: the underwriter's message, and the lead open in the page when there is one.",
        "properties": {
          "lead_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Id"
          },
          "message": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "additionalProperties": false,
        "description": "One answer. A directive the command layer accepts comes back as a card; one it refuses (an\napproval, a rejection, a send) comes back as an answer that says why, with no card (7.4, A.11).",
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "cited_event_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Cited Event Ids",
            "type": "array"
          },
          "proposal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProposalView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "answer",
          "cited_event_ids",
          "proposal"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "CommandResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "accepted",
          "event_id",
          "reason"
        ],
        "title": "CommandResponse",
        "type": "object"
      },
...
      "DraftView": {
        "additionalProperties": false,
        "description": "One of the lead's intents (A.1). A reply is accepted only for one in state `sent`.",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "intent_id": {
            "title": "Intent Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "routine_request",
              "sensitive_request",
              "quote_packet",
              "decline_notice"
            ],
            "title": "Kind",
            "type": "string"
          },
          "payload_hash": {
            "title": "Payload Hash",
            "type": "string"
          },
          "recipient": {
            "title": "Recipient",
            "type": "string"
          },
          "round": {
            "title": "Round",
            "type": "integer"
          },
          "state": {
            "enum": [
              "draft",
              "dispatching",
              "sent",
              "unknown",
              "closed_unsent"
            ],
            "title": "State",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          }
        },
        "required": [
          "intent_id",
          "payload_hash",
          "kind",
          "recipient",
          "subject",
          "body",
          "state",
          "round"
        ],
        "title": "DraftView",
        "type": "object"
      },
...
      "EventRow": {
        "additionalProperties": false,
        "description": "One of a lead's events: its id, type, actor, simulated time and a one-line summary of its payload.",
        "properties": {
          "actor": {
            "enum": [
              "workflow",
              "underwriter",
              "assistant",
              "inbound"
            ],
            "title": "Actor",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "sim_ts": {
            "title": "Sim Ts",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/EventType"
          }
        },
        "required": [
          "id",
          "type",
          "actor",
          "sim_ts",
          "summary"
        ],
        "title": "EventRow",
        "type": "object"
      },
...
      "FactField": {
        "additionalProperties": false,
        "description": "A fact key an underwriter may resolve, with the label and answer type the page offers it by.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "title": "Options",
            "type": "array"
          }
        },
        "required": [
          "key",
          "label",
          "kind",
          "options"
        ],
        "title": "FactField",
        "type": "object"
      },
      "FactView": {
        "additionalProperties": false,
        "description": "An effective fact with its source tag. `p_f` is a fact like any other.",
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "evidence": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Evidence",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "observation_id": {
            "title": "Observation Id",
            "type": "integer"
          },
          "source": {
            "enum": [
              "submitted",
              "fetched",
              "derived",
              "assumed",
              "reply",
              "underwriter"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "pending_review",
              "rejected"
            ],
            "title": "Status",
            "type": "string"
          },
          "value": {
            "$ref": "#/components/schemas/JsonValue"
          }
        },
        "required": [
          "key",
          "value",
          "source",
          "status",
          "confirmed",
          "evidence",
          "observation_id"
        ],
        "title": "FactView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
      "LeadDetail": {
        "additionalProperties": false,
        "description": "`GET /api/leads/{id}`. The rule traces are the plan's (`effects[].trace`,\n`declines_on_every_branch`); its open choices and `not_evaluated` notes are the plan's too.",
        "properties": {
          "blockers": {
            "items": {
              "$ref": "#/components/schemas/BlockerView"
            },
            "title": "Blockers",
            "type": "array"
          },
          "drafts": {
            "items": {
              "$ref": "#/components/schemas/DraftView"
            },
            "title": "Drafts",
            "type": "array"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/FactView"
            },
            "title": "Facts",
            "type": "array"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FactField"
            },
            "title": "Fields",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "lead_id": {
            "title": "Lead Id",
            "type": "string"
          },
          "plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "enum": [
              "received",
              "triaged",
              "in_progress",
              "quote_sent",
              "declined"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "lead_id",
          "label",
          "status",
          "revision",
          "facts",
          "plan",
          "blockers",
          "drafts",
          "fields"
        ],
        "title": "LeadDetail",
        "type": "object"
      },
      "LeadEvents": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventRow"
            },
            "title": "Events",
            "type": "array"
          },
          "lead_id": {
            "title": "Lead Id",
            "type": "string"
          }
        },
        "required": [
          "lead_id",
          "events"
        ],
        "title": "LeadEvents",
        "type": "object"
      },
...
      "NotEvaluatedNote": {
        "additionalProperties": false,
        "properties": {
          "producer_text": {
            "title": "Producer Text",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "text",
          "producer_text"
        ],
        "title": "NotEvaluatedNote",
        "type": "object"
      },
...
      "ProposalView": {
        "additionalProperties": false,
        "description": "A `proposals` row (A.1). A command proposal's payload is `{type, payload, rationale}`.",
        "properties": {
          "actor": {
            "enum": [
              "workflow",
              "underwriter",
              "assistant",
              "inbound"
            ],
            "title": "Actor",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "proposal_id": {
            "title": "Proposal Id",
            "type": "integer"
          },
          "state": {
            "enum": [
              "open",
              "applied",
              "dismissed"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "proposal_id",
          "payload",
          "state",
          "actor",
          "event_id"
        ],
        "title": "ProposalView",
        "type": "object"
      },
...
      "RuleTrace": {
        "additionalProperties": false,
        "description": "Either one path or a set of alternatives, never both.",
        "properties": {
          "alternatives": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TraceBranch"
            },
            "title": "Alternatives",
            "type": "array"
          },
          "board_path": {
            "items": {
              "type": "string"
            },
            "title": "Board Path",
            "type": "array"
          },
          "choice_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Choice Ids",
            "type": "array"
          }
        },
        "required": [
          "board_path",
          "alternatives",
          "choice_ids"
        ],
        "title": "RuleTrace",
        "type": "object"
      },
...
      "UndecidedPage": {
        "additionalProperties": false,
        "properties": {
          "graph": {
            "title": "Graph",
            "type": "string"
          },
          "waits_on": {
            "items": {
              "type": "string"
            },
            "title": "Waits On",
            "type": "array"
          }
        },
        "required": [
          "graph",
          "waits_on"
        ],
        "title": "UndecidedPage",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Underwriting triage",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/chat": {
      "post": {
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat"
      }
    },
...
    "/api/leads/{id}": {
      "get": {
        "operationId": "get_lead_api_leads__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lead"
      }
    },
    "/api/leads/{id}/events": {
      "get": {
        "operationId": "get_lead_events_api_leads__id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadEvents"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lead Events"
      }
    },
...
    "/api/proposals/{proposal_id}/apply": {
      "post": {
        "operationId": "apply_proposal_api_proposals__proposal_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Proposal"
      }
    },
    "/api/proposals/{proposal_id}/dismiss": {
      "post": {
        "operationId": "dismiss_proposal_api_proposals__proposal_id__dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss Proposal"
      }
    },
...
  }
}